requirements.txt
numpy==1.19.5
tensorflow==2.4.1
Pillow==8.2.0
